bot/requirements.txt
python-telegram-bot[async]